+++ src/parser/mod.rs
@@ -62,7 +62,7 @@ mod usecase;
 mod view;
 
 pub(crate) use feature_value::feature_value_part;
-pub(crate) use span::{node_from_to, span_from_to, with_span, Input};
+pub(crate) use span::{advance, node_from_to, span_from_to, with_span, Input};
 
 pub use parse::{
     parse_root, parse_root_owned, parse_with_diagnostics, parse_with_diagnostics_owned, ParseResult,